feat : 생활꿀팁 카테고리 별 조회 API 구현 완료, (Test 필요)

+++ src/main/java/umc/demoday/whatisthis/domain/post/dto/PostResponseDTO.java
@@ -48,7 +48,7 @@ public class PostResponseDTO {
         SortBy  sortBy;
         Integer page;
         Integer size;
-        Integer totalElements;
+        Long totalElements;
         Integer totalPages;
         List<PostResponseDTO.GgulPostSummaryDTO> posts;
     }
